Refactor Service creations to use Properties instead of HikariConfig
The execution of Service creations from different databases were updated to use Properties objects or properties configuration files instead of HikariConfig objects. Tests were modified accordingly and some minor code formatting changes were applied. Moreover, ConnectionPool class was created to handle transactions and improve the consistency and reusability of the code. Service class was refactored to an abstract class and took ConnectionPool and Serializer as properties which are initialized by the constructor.