Replace IsolationLevel with custom Isolation enum and adding demo snipets.
The IsolationLevel from the HikariCP library was replaced with a custom built Isolation enum to enhance database isolation control. Furthermore, the 'Benchmarks' class was renamed to 'Test_Benchmarks' to improve identification of its purpose. A variety of other services were adapted to use the new Isolation enum, and various sample test and demo files were added and modified.